feat :: RecruitmentContent 추상화
